Updated sleep data type to FLOAT in diary table

+++ docker/init.sql
@@ -3,7 +3,7 @@
 CREATE TABLE diary (
   date DATE PRIMARY KEY,
   tasks INTEGER[],
-  sleep INTEGER,
+  sleep FLOAT,
   bodybattery_min INTEGER,
   bodybattery_max INTEGER,
   steps INTEGER,
